// pleace-golf-api/src/jobs/job.enum.ts
export enum Job {
    UserEmailVerificationJob = "UserEmailVerificationJob",
    UserForgotPasswordRequestJob = "UserForgotPasswordRequestJob",
    TournamentResultJob = "TournamentResultJob",
    CreateInviteeRequestJob = "CreateInviteeRequestJob",
    NotifyClubsUpdateCourseJob = "NotifyClubsUpdateCourseJob",
    GolfClubUpdatedNotificationJob = "GolfClubUpdatedNotificationJob",
    needHelpJob = "needHelpJob"
}